AI_Resume_Screening/requirements.txt
numpy==1.26.4
scikit-learn
pdfminer.six
python-docx
streamlit
